Adds reorganization of course content component but still includes errors - adds component to show degrees courses - removes "How To" - Errors:   - when closing application modal   - when selecting course groups or course degrees
