+++ frontend-nx/apps/edu-hub/components/CourseContent/ApplicationButtonOrStatusMessageOrLinks/EnrollmentUIController/ApplicationModal/index.tsx
@@ -3,23 +3,23 @@ import Fade from '@material-ui/core/Fade';
 import Image from 'next/image';
 import { ChangeEvent, FC, useCallback, useState } from 'react';
 
-import { CourseEnrollmentStatus_enum } from '../../../../../../__generated__/globalTypes';
+import { CourseEnrollmentStatus_enum } from '../../../../../__generated__/globalTypes';
 // import { enrollmentStatusForCourse } from '../../../../../../helpers/courseHelpers';
-import { useAuthedMutation } from '../../../../../../hooks/authedMutation';
-import { useUserId } from '../../../../../../hooks/user';
-import xIcon from '../../../../../../public/images/common/x-calibur-black.svg';
-import { Course_Course_by_pk } from '../../../../../../queries/__generated__/Course';
-import { CourseWithEnrollment_Course_by_pk } from '../../../../../../queries/__generated__/CourseWithEnrollment';
+import { useAuthedMutation } from '../../../../../hooks/authedMutation';
+import { useUserId } from '../../../../../hooks/user';
+import xIcon from '../../../../../public/images/common/x-calibur-black.svg';
+import { Course_Course_by_pk } from '../../../../../queries/__generated__/Course';
+import { CourseWithEnrollment_Course_by_pk } from '../../../../../queries/__generated__/CourseWithEnrollment';
 import {
   InsertEnrollment,
   InsertEnrollmentVariables,
-} from '../../../../../../queries/__generated__/InsertEnrollment';
-import { INSERT_ENROLLMENT } from '../../../../../../queries/insertEnrollment';
+} from '../../../../../queries/__generated__/InsertEnrollment';
+import { INSERT_ENROLLMENT } from '../../../../../queries/insertEnrollment';
 
 import { CourseApplicationModalFormContent } from './FormContent';
 import { ApplicationSuccessMessage } from './ApplicationSuccessMessage';
 import { get } from 'lodash';
-import { getCourseEnrollment } from '../../../../../../helpers/util';
+import { getCourseEnrollment } from '../../../../../helpers/util';
 
 interface IProps {
   closeModal: () => void;
