+++ frontend-nx/apps/edu-hub/components/CourseContent/ApplicationButtonOrStatusMessageOrLinks/EnrollmentUIController/Links.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 import useTranslation from 'next-translate/useTranslation';
-import { Button } from '../../../../common/Button';
+import { Button } from '../../../common/Button';
 
-import { Course_Course_by_pk } from '../../../../../queries/__generated__/Course';
+import { Course_Course_by_pk } from '../../../../queries/__generated__/Course';
 interface IProps {
   course: Course_Course_by_pk;
 }
