+++ frontend-nx/apps/edu-hub/components/CourseContent/AttendancesAndAchievements/Attendances/index.tsx
@@ -1,7 +1,7 @@
 import useTranslation from 'next-translate/useTranslation';
 import { FC } from 'react';
 
-import { CourseWithEnrollment_Course_by_pk } from '../../../../../queries/__generated__/CourseWithEnrollment';
+import { CourseWithEnrollment_Course_by_pk } from '../../../../queries/__generated__/CourseWithEnrollment';
 import { BlockTitle } from '@opencampus/shared-components';
 
 import { AttendanceEntry } from './AttendanceEntry';
