--- frontend-nx/apps/edu-hub/components/CourseContent/Content/ApplicationButtonOrStatusMessageOrLinks/ApplyButton.tsx
@@ -1,56 +1,0 @@
-import useTranslation from 'next-translate/useTranslation';
-import { FC } from 'react';
-
-import { Course_Course_by_pk } from '../../../../queries/__generated__/Course';
-import { Button } from '../../../common/Button';
-
-interface IProps {
-  course: Course_Course_by_pk;
-  onClickApply: () => void;
-}
-
-// export const ApplyButtonBlock: FC<IProps> = ({ course, onClickApply }) => {
-//   const { t, lang } = useTranslation('course-page');
-//   return (
-//     <div className="flex flex-1 flex-col justify-center items-center">
-//       <Button filled onClick={onClickApply}>
-//         {t('applyNow')}
-//       </Button>
-//       <span className="text-xs mt-4">
-//         {t('applicationDeadline')}{' '}
-//         {course.applicationEnd?.toLocaleDateString(lang, {
-//           month: '2-digit',
-//           day: '2-digit',
-//         }) ?? ''}
-//       </span>
-//     </div>
-//   );
-// };
-export const ApplyButton: FC<IProps> = ({ course, onClickApply }) => {
-  const { t, lang } = useTranslation('course-page');
-  const now = new Date();
-
-  console.log(new Date());
-  console.log(course.applicationEnd);
-
-  return (
-    <div className="flex flex-1 flex-col justify-center items-center">
-      <Button
-        filled
-        inverted
-        onClick={onClickApply}
-        disabled={now > course.applicationEnd}
-        className="bg-edu-course-current"
-      >
-        {t('applyNow')}
-      </Button>
-      <span className="text-xs mt-4 text-white">
-        {t('applicationDeadline')}{' '}
-        {course.applicationEnd?.toLocaleDateString(lang, {
-          month: '2-digit',
-          day: '2-digit',
-        }) ?? ''}
-      </span>
-    </div>
-  );
-};
