+++ frontend-nx/apps/edu-hub/components/CourseContent/DescriptionFields.tsx
@@ -3,7 +3,7 @@ import useTranslation from 'next-translate/useTranslation';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 
-import { Course_Course_by_pk } from '../../../queries/__generated__/Course';
+import { Course_Course_by_pk } from '../../queries/__generated__/Course';
 
 interface IProps {
   course: Course_Course_by_pk;
