+++ frontend-nx/apps/edu-hub/components/CourseContent/TimeLocationLanguageInstructors.tsx
@@ -6,11 +6,11 @@ import {
   getEndTimeString,
   getStartTimeString,
   // getWeekdayString,
-} from '../../../helpers/dateHelpers';
-import mysteryImg from '../../../public/images/common/mystery.svg';
-import languageIcon from '../../../public/images/course/language.svg';
-import pinIcon from '../../../public/images/course/pin.svg';
-import { Course_Course_by_pk } from '../../../queries/__generated__/Course';
+} from '../../helpers/dateHelpers';
+import mysteryImg from '../../public/images/common/mystery.svg';
+import languageIcon from '../../public/images/course/language.svg';
+import pinIcon from '../../public/images/course/pin.svg';
+import { Course_Course_by_pk } from '../../queries/__generated__/Course';
 
 interface IProps {
   course: Course_Course_by_pk;
