+++ frontend-nx/apps/edu-hub/hooks/course.ts
@@ -3,5 +3,7 @@ import { CourseWithEnrollment_Course_by_pk } from "../queries/__generated__/Cour
 export const useIsCourseWithEnrollment = (
     course: any
   ): course is CourseWithEnrollment_Course_by_pk => {
+    if (!course) return false;
     return 'CourseEnrollments' in course;
   }
+
