+++ frontend-nx/apps/edu-hub/locales/de/course-page.json
@@ -2,7 +2,7 @@
   "degree_course_list_title": "Aktuelle Degree Angebote",
   "no_degree_elements_available": "Aktuell gibt es keine laufenden Angebote zu diesem Degree.",
   "degreeCourses": "Degrees",
-  "applicationDeadline": "Bewerbungsfrist",
+  "application_deadline": "Bewerbungsfrist: ",
   "applyNow": "JETZT ANMELDEN",
   "attendance": "Attendance",
   "confirm": "Confirm",
