+++ frontend-nx/apps/edu-hub/locales/en/course-page.json
@@ -1,7 +1,7 @@
 {
   "degree_course_list_title": "Current Degree Offers",
   "no_degree_elements_available": "Currently there are no ongoing offerings for this Degree.",
-  "applicationDeadline": "Application deadline",
+  "application_deadline": "Application deadline: ",
   "applyNow": "APPLY NOW",
   "attendance": "Attendance",
   "confirm": "Confirm",
